executor: create unstable test
