+++ executor/executor_test.go
@@ -6223,72 +6223,6 @@ func TestTableLockPrivilege(t *testing.T) {
 	tk.MustExec("LOCK TABLE test.t WRITE, test2.t2 WRITE")
 }
 
-func TestGlobalMemoryControl(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-
-	tk0 := testkit.NewTestKit(t, store)
-	tk0.MustExec("set global tidb_mem_oom_action = 'cancel'")
-	tk0.MustExec("set global tidb_server_memory_limit = 512 << 20")
-	tk0.MustExec("set global tidb_server_memory_limit_sess_min_size = 128")
-
-	tk1 := testkit.NewTestKit(t, store)
-	tracker1 := tk1.Session().GetSessionVars().MemTracker
-	tracker1.FallbackOldAndSetNewAction(&memory.PanicOnExceed{})
-
-	tk2 := testkit.NewTestKit(t, store)
-	tracker2 := tk2.Session().GetSessionVars().MemTracker
-	tracker2.FallbackOldAndSetNewAction(&memory.PanicOnExceed{})
-
-	tk3 := testkit.NewTestKit(t, store)
-	tracker3 := tk3.Session().GetSessionVars().MemTracker
-	tracker3.FallbackOldAndSetNewAction(&memory.PanicOnExceed{})
-
-	sm := &testkit.MockSessionManager{
-		PS: []*util.ProcessInfo{tk1.Session().ShowProcess(), tk2.Session().ShowProcess(), tk3.Session().ShowProcess()},
-	}
-	dom.ServerMemoryLimitHandle().SetSessionManager(sm)
-	go dom.ServerMemoryLimitHandle().Run()
-
-	tracker1.Consume(100 << 20) // 100 MB
-	tracker2.Consume(200 << 20) // 200 MB
-	tracker3.Consume(300 << 20) // 300 MB
-
-	test := make([]int, 128<<20)       // Keep 1GB HeapInUse
-	time.Sleep(500 * time.Millisecond) // The check goroutine checks the memory usage every 100ms. The Sleep() make sure that Top1Tracker can be Canceled.
-
-	// Kill Top1
-	require.False(t, tracker1.NeedKill.Load())
-	require.False(t, tracker2.NeedKill.Load())
-	require.True(t, tracker3.NeedKill.Load())
-	require.Equal(t, memory.MemUsageTop1Tracker.Load(), tracker3)
-	util.WithRecovery( // Next Consume() will panic and cancel the SQL
-		func() {
-			tracker3.Consume(1)
-		}, func(r interface{}) {
-			require.True(t, strings.Contains(r.(string), "Out Of Memory Quota!"))
-		})
-	tracker2.Consume(300 << 20) // Sum 500MB, Not Panic, Waiting t3 cancel finish.
-	time.Sleep(500 * time.Millisecond)
-	require.False(t, tracker2.NeedKill.Load())
-	// Kill Finished
-	tracker3.Consume(-(300 << 20))
-	// Simulated SQL is Canceled and the time is updated
-	sm.PSMu.Lock()
-	ps := *sm.PS[2]
-	ps.Time = time.Now()
-	sm.PS[2] = &ps
-	sm.PSMu.Unlock()
-	time.Sleep(500 * time.Millisecond)
-	// Kill the Next SQL
-	util.WithRecovery( // Next Consume() will panic and cancel the SQL
-		func() {
-			tracker2.Consume(1)
-		}, func(r interface{}) {
-			require.True(t, strings.Contains(r.(string), "Out Of Memory Quota!"))
-		})
-	require.Equal(t, test[0], 0) // Keep 1GB HeapInUse
-}
-
 func TestGlobalMemoryControl2(t *testing.T) {
 	store, dom := testkit.CreateMockStoreAndDomain(t)
