+++ executor/unstabletest/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "unstabletest_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "unstable_test.go",
+    ],
+    flaky = True,
+    race = "on",
+    shard_count = 2,
+    deps = [
+        "//config",
+        "//meta/autoid",
+        "//testkit",
+        "//util",
+        "//util/memory",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@io_opencensus_go//stats/view",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
