+++ executor/unstabletest/README.md
@@ -1,0 +1,2 @@
+All unstable tests are saved here. Please fix it as soon as possible. But don't skip it or delete it so that we will lose 
+valuable test cases.
